refactor: move SpinalSimBmbInterconnectGeneratorTester
