+++ tester/src/test/scala/spinal/tester/GenerationShould.scala
@@ -1,16 +1,8 @@
 package spinal.tester
 
-import java.io.File
-import org.apache.commons.io.FileUtils
-
 import org.scalatest.funsuite.AnyFunSuite
 
 import spinal.core._
-import spinal.core.internals.GraphUtils
-
-import spinal.lib.com.i2c._
-import spinal.lib.com.uart.{UartCtrl, UartCtrlGenerics}
-import spinal.lib.soc.pinsec.{Pinsec, PinsecConfig}
 
 class GenerationShouldTester extends AnyFunSuite {
   test("BlackBoxInputUnconnected") {
@@ -457,63 +449,6 @@ class GenerationShouldTester extends AnyFunSuite {
 //    assert(ScopeProperty.get.isEmpty)
   }
 
-}
-
-class RepeatabilityTester extends AnyFunSuite {
-  var checkOutputHashCounter = 0
-  def checkOutputHash(gen: => Component): Unit = {
-    checkOutputHashCounter = checkOutputHashCounter + 1
-    var ref = ""
-    for (i <- 0 until 8) {
-      val report =
-        SpinalConfig(defaultClockDomainFrequency = FixedFrequency(50 MHz))
-          .generateVerilog(
-            gen.setDefinitionName(s"checkOutputHash_${checkOutputHashCounter}")
-          )
-      FileUtils.copyFile(
-        new File(report.generatedSourcesPaths.head),
-        new File(report.generatedSourcesPaths.head + "_" + i + ".v")
-      )
-
-      import sys.process._
-      val hash = s"md5sum ${report.generatedSourcesPaths.head}".!!
-      if (i == 0) ref = hash
-      else assert(ref == hash)
-    }
-  }
-  def configI2C = I2cSlaveMemoryMappedGenerics(
-    ctrlGenerics = I2cSlaveGenerics(),
-    addressFilterCount = 0,
-    masterGenerics = I2cMasterMemoryMappedGenerics(timerWidth = 32)
-  )
-
-  test("Apb3I2cCtrlGraph") {
-    val dut = SpinalConfig(defaultClockDomainFrequency = FixedFrequency(50 MHz))
-      .generateVerilog(new Apb3I2cCtrl(configI2C))
-      .toplevel
-    assert(GraphUtils.countNames(dut) == 221)
-  }
-
-  test("UartGraph") {
-    val dut = SpinalVerilog(new UartCtrl(UartCtrlGenerics())).toplevel
-    assert(GraphUtils.countNames(dut) == 94)
-  }
-
-  test("Apb3I2cCtrlVerilog") {
-    checkOutputHash(new Apb3I2cCtrl(configI2C))
-  }
-
-  test("UartVerilog") {
-    checkOutputHash(new UartCtrl(UartCtrlGenerics()))
-  }
-
-  test("PinsecVerilog") {
-    checkOutputHash(new Pinsec(PinsecConfig.default))
-  }
-
-  test("BmbInterconnectVerilog") {
-    checkOutputHash(scalatest.SpinalSimBmbInterconnectGeneratorTester.component)
-  }
 }
 
 class NameingTester extends AnyFunSuite {
